@@ -1,16 +1,5 @@
-module github.com/fasthttp/router
+module github.com/sirkostya009/httx
 
-go 1.21
+go 1.22
 
 toolchain go1.23.4
-
-require (
-	github.com/savsgio/gotils v0.0.0-20240704082632-aef3928b8a38
-	github.com/valyala/bytebufferpool v1.0.0
-	github.com/valyala/fasthttp v1.58.0
-)
-
-require (
-	github.com/andybalholm/brotli v1.1.1 // indirect
-	github.com/klauspost/compress v1.17.11 // indirect
-)
